--- a/src/lexer/state.rs
+++ b/src/lexer/state.rs
@@ -3,7 +3,7 @@ use std::collections::VecDeque;
 use crate::lexer::byte_string::ByteString;
 use crate::lexer::error::SyntaxError;
 use crate::lexer::error::SyntaxResult;
-use crate::lexer::token::Span;
+use crate::lexer::source::Source;
 
 #[derive(Debug, PartialEq, Eq, PartialOrd, Clone, Copy)]
 pub enum DocStringKind {
@@ -11,7 +11,7 @@ pub enum DocStringKind {
     Nowdoc,
 }
 
-#[derive(Debug, PartialEq, Eq, Clone)]
+#[derive(Debug)]
 pub enum StackFrame {
     Initial,
     Scripting,
@@ -24,36 +24,24 @@ pub enum StackFrame {
     VarOffset,
 }
 
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct State {
+#[derive(Debug)]
+pub struct State<'a> {
+    pub source: Source<'a>,
     pub stack: VecDeque<StackFrame>,
-    pub chars: Vec<u8>,
-    pub cursor: usize,
-    pub current: Option<u8>,
-    pub span: Span,
 }
 
-impl State {
-    pub fn new<B: ?Sized + AsRef<[u8]>>(input: &B) -> Self {
-        let chars = input.as_ref().to_vec();
-        let current = chars.first().copied();
-
-        let mut stack = VecDeque::with_capacity(32);
-        stack.push_back(StackFrame::Initial);
-
+impl<'a> State<'a> {
+    pub fn new(source: Source<'a>) -> Self {
         Self {
-            stack,
-            chars,
-            current,
-            cursor: 0,
-            span: (1, 1),
+            source,
+            stack: VecDeque::from([StackFrame::Initial]),
         }
     }
 
     pub fn frame(&self) -> SyntaxResult<&StackFrame> {
         self.stack
             .back()
-            .ok_or(SyntaxError::UnpredictableState(self.span))
+            .ok_or_else(|| SyntaxError::UnpredictableState(self.source.span()))
     }
 
     pub fn replace(&mut self, state: StackFrame) {
@@ -68,36 +56,5 @@ impl State {
 
     pub fn exit(&mut self) {
         self.stack.pop_back();
-    }
-
-    pub fn peek_buf(&self) -> &[u8] {
-        &self.chars[self.cursor..]
-    }
-
-    pub fn peek_byte(&self, delta: usize) -> Option<u8> {
-        self.chars.get(self.cursor + delta).copied()
-    }
-
-    pub fn try_read(&self, search: &[u8]) -> bool {
-        self.peek_buf().starts_with(search)
-    }
-
-    pub fn skip(&mut self, count: usize) {
-        for _ in 0..count {
-            self.next();
-        }
-    }
-
-    pub fn next(&mut self) {
-        match self.current {
-            Some(b'\n') => {
-                self.span.0 += 1;
-                self.span.1 = 1;
-            }
-            Some(_) => self.span.1 += 1,
-            _ => {}
-        }
-        self.cursor += 1;
-        self.current = self.chars.get(self.cursor).copied();
     }
 }