--- a/tests/fixtures/0269/ast.txt
+++ b/tests/fixtures/0269/ast.txt
@@ -1,0 +1,504 @@
+[
+    Constant {
+        constants: [
+            Constant {
+                name: Identifier {
+                    start: (
+                        3,
+                        7,
+                    ),
+                    name: "\xe2\x86\xaa",
+                    end: (
+                        3,
+                        11,
+                    ),
+                },
+                value: LiteralString {
+                    value: "\n",
+                },
+            },
+        ],
+    },
+    Interface {
+        name: Identifier {
+            start: (
+                5,
+                11,
+            ),
+            name: "\xf0\x9f\x94\x8a",
+            end: (
+                5,
+                16,
+            ),
+        },
+        attributes: [],
+        extends: [],
+        body: [
+            Method(
+                Method {
+                    start: (
+                        6,
+                        3,
+                    ),
+                    end: (
+                        6,
+                        43,
+                    ),
+                    name: Identifier {
+                        start: (
+                            6,
+                            19,
+                        ),
+                        name: "\xf0\x9f\x93\x9d",
+                        end: (
+                            6,
+                            23,
+                        ),
+                    },
+                    attributes: [],
+                    parameters: MethodParameterList {
+                        start: (
+                            6,
+                            23,
+                        ),
+                        end: (
+                            6,
+                            37,
+                        ),
+                        members: [
+                            MethodParameter {
+                                start: (
+                                    6,
+                                    24,
+                                ),
+                                end: (
+                                    6,
+                                    36,
+                                ),
+                                name: Variable {
+                                    start: (
+                                        6,
+                                        31,
+                                    ),
+                                    name: "\xf0\x9f\x93\x83",
+                                    end: (
+                                        6,
+                                        36,
+                                    ),
+                                },
+                                attributes: [],
+                                type: Some(
+                                    String,
+                                ),
+                                variadic: false,
+                                default: None,
+                                modifiers: PromotedPropertyModifierGroup {
+                                    modifiers: [],
+                                },
+                                by_ref: false,
+                            },
+                        ],
+                    },
+                    body: None,
+                    modifiers: MethodModifierGroup {
+                        modifiers: [
+                            Public {
+                                start: (
+                                    6,
+                                    3,
+                                ),
+                                end: (
+                                    6,
+                                    10,
+                                ),
+                            },
+                        ],
+                    },
+                    return_type: Some(
+                        Void,
+                    ),
+                    by_ref: false,
+                },
+            ),
+        ],
+    },
+    Class {
+        name: Identifier {
+            start: (
+                9,
+                13,
+            ),
+            name: "\xf0\x9f\x92\xbb",
+            end: (
+                9,
+                18,
+            ),
+        },
+        attributes: [],
+        extends: None,
+        implements: [
+            Identifier {
+                start: (
+                    9,
+                    29,
+                ),
+                name: "\xf0\x9f\x94\x8a",
+                end: (
+                    9,
+                    34,
+                ),
+            },
+        ],
+        body: [
+            Method(
+                Method {
+                    start: (
+                        10,
+                        3,
+                    ),
+                    end: (
+                        12,
+                        3,
+                    ),
+                    name: Identifier {
+                        start: (
+                            10,
+                            19,
+                        ),
+                        name: "\xf0\x9f\x93\x9d",
+                        end: (
+                            10,
+                            23,
+                        ),
+                    },
+                    attributes: [],
+                    parameters: MethodParameterList {
+                        start: (
+                            10,
+                            23,
+                        ),
+                        end: (
+                            10,
+                            37,
+                        ),
+                        members: [
+                            MethodParameter {
+                                start: (
+                                    10,
+                                    24,
+                                ),
+                                end: (
+                                    10,
+                                    36,
+                                ),
+                                name: Variable {
+                                    start: (
+                                        10,
+                                        31,
+                                    ),
+                                    name: "\xf0\x9f\x93\x83",
+                                    end: (
+                                        10,
+                                        36,
+                                    ),
+                                },
+                                attributes: [],
+                                type: Some(
+                                    String,
+                                ),
+                                variadic: false,
+                                default: None,
+                                modifiers: PromotedPropertyModifierGroup {
+                                    modifiers: [],
+                                },
+                                by_ref: false,
+                            },
+                        ],
+                    },
+                    body: Some(
+                        [
+                            Expression {
+                                expr: Print {
+                                    value: Infix {
+                                        lhs: Variable(
+                                            Variable {
+                                                start: (
+                                                    11,
+                                                    11,
+                                                ),
+                                                name: "\xf0\x9f\x93\x83",
+                                                end: (
+                                                    11,
+                                                    17,
+                                                ),
+                                            },
+                                        ),
+                                        op: Concat,
+                                        rhs: Identifier(
+                                            Identifier {
+                                                start: (
+                                                    11,
+                                                    19,
+                                                ),
+                                                name: "\xe2\x86\xaa",
+                                                end: (
+                                                    11,
+                                                    22,
+                                                ),
+                                            },
+                                        ),
+                                    },
+                                },
+                            },
+                        ],
+                    ),
+                    modifiers: MethodModifierGroup {
+                        modifiers: [
+                            Public {
+                                start: (
+                                    10,
+                                    3,
+                                ),
+                                end: (
+                                    10,
+                                    10,
+                                ),
+                            },
+                        ],
+                    },
+                    return_type: Some(
+                        Void,
+                    ),
+                    by_ref: false,
+                },
+            ),
+        ],
+        modifiers: ClassModifierGroup {
+            modifiers: [
+                Final {
+                    start: (
+                        9,
+                        1,
+                    ),
+                    end: (
+                        9,
+                        7,
+                    ),
+                },
+            ],
+        },
+    },
+    Function(
+        Function {
+            start: (
+                15,
+                1,
+            ),
+            end: (
+                19,
+                1,
+            ),
+            name: Identifier {
+                start: (
+                    15,
+                    10,
+                ),
+                name: "\xf0\x9f\x9a\xaa",
+                end: (
+                    15,
+                    14,
+                ),
+            },
+            attributes: [],
+            parameters: FunctionParameterList {
+                start: (
+                    15,
+                    14,
+                ),
+                end: (
+                    15,
+                    16,
+                ),
+                members: [],
+            },
+            return_type: Some(
+                Void,
+            ),
+            by_ref: false,
+            body: [
+                Expression {
+                    expr: Infix {
+                        lhs: Variable(
+                            Variable {
+                                start: (
+                                    16,
+                                    3,
+                                ),
+                                name: "\xf0\x9f\x92\xbb",
+                                end: (
+                                    16,
+                                    9,
+                                ),
+                            },
+                        ),
+                        op: Assign,
+                        rhs: New {
+                            target: Identifier(
+                                Identifier {
+                                    start: (
+                                        16,
+                                        15,
+                                    ),
+                                    name: "\xf0\x9f\x92\xbb",
+                                    end: (
+                                        16,
+                                        19,
+                                    ),
+                                },
+                            ),
+                            args: [],
+                        },
+                    },
+                },
+                Expression {
+                    expr: Infix {
+                        lhs: Variable(
+                            Variable {
+                                start: (
+                                    17,
+                                    3,
+                                ),
+                                name: "\xf0\x9f\x93\x84",
+                                end: (
+                                    17,
+                                    9,
+                                ),
+                            },
+                        ),
+                        op: Assign,
+                        rhs: LiteralString {
+                            value: "hello, world",
+                        },
+                    },
+                },
+                Expression {
+                    expr: MethodCall {
+                        target: Variable(
+                            Variable {
+                                start: (
+                                    18,
+                                    3,
+                                ),
+                                name: "\xf0\x9f\x92\xbb",
+                                end: (
+                                    18,
+                                    8,
+                                ),
+                            },
+                        ),
+                        method: Identifier(
+                            Identifier {
+                                start: (
+                                    18,
+                                    10,
+                                ),
+                                name: "\xf0\x9f\x93\x9d",
+                                end: (
+                                    18,
+                                    14,
+                                ),
+                            },
+                        ),
+                        args: [
+                            Arg {
+                                name: Some(
+                                    Identifier {
+                                        start: (
+                                            18,
+                                            15,
+                                        ),
+                                        name: "\xf0\x9f\x93\x83",
+                                        end: (
+                                            18,
+                                            19,
+                                        ),
+                                    },
+                                ),
+                                value: Variable(
+                                    Variable {
+                                        start: (
+                                            18,
+                                            21,
+                                        ),
+                                        name: "\xf0\x9f\x93\x84",
+                                        end: (
+                                            18,
+                                            26,
+                                        ),
+                                    },
+                                ),
+                                unpack: false,
+                            },
+                        ],
+                    },
+                },
+            ],
+        },
+    ),
+    Expression {
+        expr: Call {
+            target: Identifier(
+                Identifier {
+                    start: (
+                        21,
+                        1,
+                    ),
+                    name: "\xf0\x9f\x9a\xaa",
+                    end: (
+                        21,
+                        5,
+                    ),
+                },
+            ),
+            args: [],
+        },
+    },
+    Expression {
+        expr: Infix {
+            lhs: Variable(
+                Variable {
+                    start: (
+                        23,
+                        1,
+                    ),
+                    name: "var\xef\xbf\xbd",
+                    end: (
+                        23,
+                        9,
+                    ),
+                },
+            ),
+            op: Assign,
+            rhs: LiteralInteger {
+                i: "1",
+            },
+        },
+    },
+    Echo {
+        values: [
+            Variable(
+                Variable {
+                    start: (
+                        23,
+                        19,
+                    ),
+                    name: "var\xef\xbf\xbd",
+                    end: (
+                        23,
+                        26,
+                    ),
+                },
+            ),
+        ],
+    },
+]