--- a/tests/fixtures/0269/code.php
+++ b/tests/fixtures/0269/code.php
@@ -1,0 +1,23 @@
+<?php
+
+const ↪ = "\n";
+
+interface 🔊 {
+  public function 📝(string $📃): void;
+}
+
+final class 💻 implements 🔊 {
+  public function 📝(string $📃): void {
+    print($📃 . ↪);
+  }
+}
+
+function 🚪(): void {
+  $💻 = new 💻();
+  $📄 = "hello, world";
+  $💻->📝(📃: $📄);
+}
+
+🚪();
+
+$var� = 1; echo $var�;