… src/lexer/mod.rs
… tests/fixtures/0269/tokens.txt