Not much of a performance gain, but it makes lexer state easier to work with.

benchmarks:

before:

```
  Time (mean ± σ):      2.519 s ±  0.017 s    [User: 2.611 s, System: 0.225 s]
  Range (min … max):    2.501 s …  2.551 s    10 runs
```

after:

```
  Time (mean ± σ):      2.451 s ±  0.007 s    [User: 2.561 s, System: 0.190 s]
  Range (min … max):    2.443 s …  2.462 s    10 runs
```